sluggram-vite/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
cloudinary==1.38.0
